Increased debug and possible bugfix
