+++ apps/muddle-node/main.cpp
@@ -23,8 +23,8 @@
 #include "core/macros.hpp"
 #include "core/synchronisation/protected.hpp"
 #include "crypto/ecdsa.hpp"
+#include "http/http_server.hpp"
 #include "http/module.hpp"
-#include "http/server.hpp"
 #include "muddle/muddle_endpoint.hpp"
 #include "muddle/muddle_interface.hpp"
 #include "network/management/network_manager.hpp"
