+++ apps/muddle-test-network/main.cpp
@@ -25,10 +25,10 @@
 #include "crypto/ecdsa.hpp"
 #include "crypto/prover.hpp"
 #include "crypto/sha256.hpp"
+#include "http/http_server.hpp"
 #include "http/middleware/allow_origin.hpp"
 #include "http/middleware/telemetry.hpp"
 #include "http/module.hpp"
-#include "http/server.hpp"
 #include "muddle/muddle_endpoint.hpp"
 #include "muddle/muddle_interface.hpp"
 #include "network/management/network_manager.hpp"
