+++ libs/constellation/include/constellation/constellation.hpp
@@ -20,8 +20,8 @@
 #include "core/feature_flags.hpp"
 #include "core/reactor.hpp"
 #include "entropy/entropy_generator_interface.hpp"
+#include "http/http_server.hpp"
 #include "http/module.hpp"
-#include "http/server.hpp"
 #include "ledger/block_sink_interface.hpp"
 #include "ledger/chain/block_coordinator.hpp"
 #include "ledger/chain/consensus/consensus_miner_interface.hpp"
