+++ libs/http/examples/server/server.cpp
@@ -16,10 +16,10 @@
 //
 //------------------------------------------------------------------------------
 
+#include "http/http_server.hpp"
 #include "http/json_response.hpp"
 #include "http/middleware/deny_all.hpp"
 #include "http/middleware/token_auth.hpp"
-#include "http/server.hpp"
 #include "http/validators.hpp"
 #include "json/document.hpp"
