+++ libs/http/include/http/middleware/telemetry.hpp
@@ -17,7 +17,7 @@
 //
 //------------------------------------------------------------------------------
 
-#include "http/server.hpp"
+#include "http/http_server.hpp"
 
 namespace fetch {
 namespace http {
