+++ libs/http/tests/unit/destruction_test.cpp
@@ -16,9 +16,9 @@
 //
 //------------------------------------------------------------------------------
 
+#include "http/http_server.hpp"
 #include "http/json_client.hpp"
 #include "http/json_response.hpp"
-#include "http/server.hpp"
 #include "json/document.hpp"
 #include "network/management/network_manager.hpp"
