+++ libs/messenger/tests/unittests/shared_functions.hpp
@@ -18,8 +18,8 @@
 //------------------------------------------------------------------------------
 
 #include "crypto/ecdsa.hpp"
+#include "http/http_server.hpp"
 #include "http/json_client.hpp"
-#include "http/server.hpp"
 #include "messenger/messenger_api.hpp"
 #include "messenger/messenger_http_interface.hpp"
 #include "messenger/messenger_prototype.hpp"
